src/backend/requirements.txt
fastapi
pydantic
supabase
python-dotenv
uvicorn[standard]